+++ trae_agent/utils/openai_compatible/providers/__init__.py
@@ -1,0 +1,14 @@
+# Copyright (c) 2025 ByteDance Ltd. and/or its affiliates
+# SPDX-License-Identifier: MIT
+
+"""Provider-specific configurations for OpenAI-compatible clients."""
+
+from .azure import AzureProvider
+from .doubao import DoubaoProvider
+from .openrouter import OpenRouterProvider
+
+__all__ = [
+    "OpenRouterProvider",
+    "AzureProvider",
+    "DoubaoProvider",
+]
